requirements.txt
cairocffi==0.7.2
CairoSVG==1.0.22
cairotft==0.1.2
cffi==1.7.0
cssselect==0.9.2
Django==1.9.6
django-bootstrap3==7.0.1
django-pagedown==0.1.1
freeze==1.0.10
html5lib==0.9999999
lxml==3.6.0
Markdown==2.6.6
mistune==0.7.3
oauthlib==1.1.2
pycparser==2.14
Pygments==2.1.3
PyJWT==1.4.1
Pyphen==0.9.4
python-openid==2.2.5
python-social-auth==0.2.19
requests==2.10.0
requests-oauthlib==0.6.2
six==1.10.0
sorl-thumbnail==12.3
tinycss==0.3
WeasyPrint==0.29
webencodings==0.5
